test_sort.c

+++ insert_sort.c
@@ -1,8 +1,10 @@
+#include "insert_sort.h"
+
 void _insert_sort_incremental (int cards[], size_t card_num)
 {
 	for (size_t i = 1; i <= card_num - 1; i++){
 		int cur = cards[i];
-		size_t j;
+		long long j;
 		for (j = i - 1; j >= 0 && cards[j] > cur; j--){
 			cards[j + 1] = cards[j];
 		}
